Added 'skip' action
